initial commit for adding unit tests for AllocSupport class
